+++ web/app/containers/app/actions.js
@@ -25,7 +25,7 @@ import * as checkoutConfirmationActions from '../checkout-confirmation/actions'
 import * as checkoutShippingActions from '../../store/checkout/shipping/actions'
 import * as cartActions from '../../store/cart/actions'
 import * as loginActions from '../login/actions'
-import * as productDetailsActions from '../product-details/actions'
+import {process as processProductDetails} from '../../integration-manager/_merlins-connector/product-details/parsers'
 import * as footerActions from '../footer/actions'
 import * as navigationActions from '../navigation/actions'
 import * as productsActions from '../../store/products/actions'
@@ -142,9 +142,6 @@ export const fetchPage = (url, pageComponent, routeName, fetchUrl) => {
 
                 if (pageComponent === Login) {
                     dispatch(loginActions.process(receivedAction))
-                } else if (pageComponent === ProductDetails) {
-                    dispatch(productDetailsActions.process(receivedAction))
-                    // dispatch(productsActions.processProductDetails(receivedAction))
                 } else if (pageComponent === ProductList) {
                     dispatch(categoriesActions.process(receivedAction))
                     dispatch(productsActions.processProductList(receivedAction))
