+++ web/app/integration-manager/_merlins-connector/product-details/commands.js
@@ -1,17 +1,22 @@
 import {urlToPathKey} from '../../../utils/utils'
 import {receiveFormInfo} from '../../actions'
-import {receivePdpProductData} from '../../responses'
 
 import {fetchPageData} from '../utils'
 
-import {productDetailsParser, pdpAddToCartFormParser} from './parsers'
+import {receiveProductDetailsProductData, receiveProductDetailsUIData} from '../../product-details/responses'
+import {productDetailsParser, productDetailsUIParser, pdpAddToCartFormParser} from './parsers'
 
 export const fetchPdpData = (url) => (dispatch) => {
     return dispatch(fetchPageData(url))
         .then((res) => {
             const [$, $response] = res
-            dispatch(receivePdpProductData({[urlToPathKey(url)]: productDetailsParser($, $response)}))
-            dispatch(receiveFormInfo({[urlToPathKey(url)]: pdpAddToCartFormParser($, $response).formInfo}))
+
+            const pathKey = urlToPathKey(url)
+
+            debugger
+            dispatch(receiveProductDetailsUIData({[pathKey]: productDetailsUIParser($, $response)}))
+            dispatch(receiveProductDetailsProductData({[pathKey]: productDetailsParser($, $response)}))
+            dispatch(receiveFormInfo({[pathKey]: pdpAddToCartFormParser($, $response).formInfo}))
         })
         .catch((error) => { console.info(error.message) })
 }
