+++ web/app/integration-manager/_merlins-connector/product-details/parsers.test.js
@@ -1,7 +1,7 @@
 /* eslint-env jquery, jest, node */
 import {jquerifyHtmlFile} from 'progressive-web-sdk/dist/test-utils'
 import {isURL} from 'validator'
-import {productDetailsParser} from './parsers'
+import {productDetailsParser, productDetailsUIParser} from './parsers'
 
 /* eslint-disable max-nested-callbacks */
 
@@ -32,5 +32,18 @@ describe('the ProductDetails product parser', () => {
 
     it('extracts the description from the page', () => {
         expect(typeof parsedContent.description).toBe('string')
+    })
+})
+
+describe('the ProductDetails UI parser', () => {
+    const $content = jquerifyHtmlFile(`${__dirname}/product-details-example.html`)
+    const parsedContent = productDetailsUIParser($, $content)
+
+    test('extracts form info from the add-to-cart form', () => {
+        expect(isURL(parsedContent.formInfo.submitUrl)).toBe(true)
+        expect(parsedContent.formInfo.method).toBe('post')
+        Object.keys(parsedContent.formInfo.hiddenInputs).forEach((key) => {
+            expect(typeof parsedContent.formInfo.hiddenInputs[key]).toBe('string')
+        })
     })
 })
