+++ web/app/integration-manager/product-details/responses.js
@@ -1,0 +1,6 @@
+import {createAction} from '../../utils/utils'
+
+export const receiveProductDetailsUIData = createAction('Receive Product Details UI data')
+
+export const receiveProductDetailsProductData = createAction('Receive Product Details product data')
+
