+++ web/app/store/products/reducer.js
@@ -1,13 +1,13 @@
 import Immutable from 'immutable'
 import {handleActions} from 'redux-actions'
 import {mergePayloadForActions} from '../../utils/reducer-utils'
-import {receivePdpProductData} from '../../integration-manager/responses'
+import {receiveProductDetailsProductData} from '../../integration-manager/product-details/responses'
 import {receiveProductListProductData} from './actions'
 
 const initialState = Immutable.Map()
 
 const productReducer = handleActions({
-    ...mergePayloadForActions(receivePdpProductData),
+    ...mergePayloadForActions(receiveProductDetailsProductData),
     [receiveProductListProductData]: (state, {payload}) => {
         return state.mergeDeepWith((prev, next) => prev || next, payload)
     }
